requirements_runtime_stac.txt
affine==2.4.0
boto3==1.42.66
botocore==1.42.66
numpy==2.0.2
pg8000==1.31.5
PyWavelets==1.6.0
pyproj==3.6.1
python-dotenv==1.2.1
PyYAML==6.0.2
rasterio==1.4.3
requests==2.32.5
scipy==1.13.1
shapely==2.0.7
torchinfo==1.8.0
torchvision==0.23.0
tqdm==4.67.1
pytorch-wavelets==1.3.0
